image_app/images/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cleanup==7.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
easy-thumbnails==2.8.5
gunicorn==20.1.0
numpy==1.24.2
opencv-python-headless
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7

